Refactor XBRL mapping
- Refactor XBRL mapping and processing: update imports, add serializers, and implement standardized response utilities
