+++ xbrl_mapping/services/pydantic_models.py
